Code Rework
Trying to make the ships move, an unpredicted error related to textures popped up and were unable to fix it.

Testing, we found a different sample of premade code (handouts) were we managed to make it work.

We figured it would take less work to port our features to a new base than debugg the original code alltogether, so this is the first instance of the new code which we will be using from now on.